+++ src/main/java/com/daduo/api/tiktokapi/model/CreditRequest.java
@@ -8,9 +8,9 @@ import lombok.Data;
 @ApiModel(value = "充值币积分请求Json")
 public class CreditRequest {
     @ApiModelProperty(value = "充值币")
-    private Long credit;
+    private Double credit;
     @ApiModelProperty(value = "用户ID")
     private Long userId;
     @ApiModelProperty(value = "积分数")
-    private Long points;
+    private Double points;
 }
